VertexShader DirectX로 변경, HLSL D3DCompile()로 처리

+++ GameEngine/GameEngineRenderTarget.cpp
@@ -49,8 +49,8 @@ void GameEngineRenderTarget::Setting(int _Index)
 
 	if (-1 == _Index)
 	{
-		GameEngineDirectXDevice::GetContext()->OMSetRenderTargets(RenderTargetViews_.size(),
-			&RenderTargetViews_[0], nullptr);
+		GameEngineDirectXDevice::GetContext()->OMSetRenderTargets(
+			static_cast<UINT>(RenderTargetViews_.size()), &RenderTargetViews_[0], nullptr);
 	}
 	else
 	{
